# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AfriChart EMR API",
        "description": "REST API for the AfriChart EMR — patients, consultations, prescriptions, invoices and the patient queue. All endpoints (except login) require a Sanctum bearer token.",
        "contact": {
            "name": "Lymora Tech",
            "email": "michael@mgbah.dev"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate and receive an API token",
                "operationId": "985cd5203e2d0a820e113f2f4fdb2847",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "doctor@africhart.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated — returns user + token"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke the current API token",
                "operationId": "a05022a34eef599e04fb0a5254375b84",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated user",
                "operationId": "5df1e3769ba9fc9f727bcf278c51470f",
                "responses": {
                    "200": {
                        "description": "Current user"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/consultations": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "List consultations (paginated)",
                "operationId": "49b937dc81f039aed05a57bbdb3a4daf",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "in_progress",
                                "completed",
                                "follow_up"
                            ]
                        }
                    },
                    {
                        "name": "mine",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated consultation list"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Start a consultation",
                "operationId": "873ca7aeffa41b854a6b4d5a38339150",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "chief_complaint",
                                    "clinical_notes"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "chief_complaint": {
                                        "type": "string"
                                    },
                                    "clinical_notes": {
                                        "type": "string"
                                    },
                                    "diagnosis": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "plan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consultation started"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/consultations/{consultation}": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Show a consultation (with prescriptions + invoice)",
                "operationId": "aaebe74f4ee179a78a2ef9dc4ac10f31",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consultation detail"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Update consultation notes (owner doctor or admin)",
                "operationId": "6f84ee032c19c687d1bdb19e31188455",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/consultations/{consultation}/vitals": {
            "patch": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Record vitals (doctor, nurse or admin)",
                "operationId": "c2d88e2c4778ea9cbd66b7d96b66803c",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "temperature": {
                                        "type": "number",
                                        "example": 37.2
                                    },
                                    "blood_pressure": {
                                        "type": "string",
                                        "example": "120/80"
                                    },
                                    "pulse_rate": {
                                        "type": "integer",
                                        "example": 72
                                    },
                                    "weight": {
                                        "type": "number",
                                        "example": 68
                                    },
                                    "height": {
                                        "type": "number",
                                        "example": 170
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vitals recorded"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/consultations/{consultation}/complete": {
            "patch": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Mark a consultation completed",
                "operationId": "1248db21c0b9f4b4f7188685aebf67f3",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Completed"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token returned by POST /auth/login.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}